+++ src/cards/components/sensor-collection.ts
@@ -6,7 +6,8 @@ import type { HomeAssistant } from '@hass/types';
 import { stateDisplay } from '@html/state-display';
 import { sensorStyles } from '@theme/styles';
 import { stylesToHostCss } from '@theme/util/style-converter';
-import type { Config, SensorData } from '@type/config';
+import type { Config } from '@type/config';
+import type { SensorData } from '@type/sensor';
 import { CSSResult, LitElement, html, nothing, type TemplateResult } from 'lit';
 import { property, state } from 'lit/decorators.js';
 import { until } from 'lit/directives/until.js';
