+++ src/delegates/entities/icon-entities.ts
@@ -1,7 +1,8 @@
 import { hasFeature } from '@config/feature';
 import { getState } from '@delegates/retrievers/state';
 import type { HomeAssistant } from '@hass/types';
-import type { Config, EntityConfig, EntityInformation } from '@type/config';
+import type { Config, EntityConfig } from '@type/config';
+import type { EntityInformation } from '@type/room';
 
 const climateIcons = {
   auto: 'mdi:autorenew',
