+++ src/types/config.ts
@@ -113,81 +113,3 @@ export interface EntityConfig {
   /** Action to perform on double tap */
   double_tap_action?: ActionConfig;
 }
-
-/**
- * Combined entity configuration and state information.
- */
-export interface EntityInformation {
-  /** The entity configuration */
-  config: EntityConfig;
-
-  /** The entity state */
-  state: EntityState | undefined;
-}
-
-export interface EntityState {
-  /** ID of the entity this state belongs to */
-  entity_id: string;
-
-  /** Current state value as a string (e.g., "on", "off", "25.5") */
-  state: string;
-
-  /** Additional attributes associated with the state */
-  attributes: Record<string, any>;
-
-  /** Returns the domain portion of the entity_id */
-  domain: string;
-}
-
-export interface RoomInformation {
-  /** The area name */
-  area_name: string;
-}
-
-/**
- * Represents an averaged sensor reading for a specific device class
- */
-export interface AveragedSensor {
-  /** Array of entity states that were averaged */
-  states: EntityState[];
-
-  /** Unit of measurement */
-  uom: string;
-
-  /** Averaged value */
-  average: number;
-
-  /** Device class (e.g., 'temperature', 'humidity') */
-  device_class: string;
-
-  /** Sensor domain (typically 'sensor') */
-  domain: string;
-}
-
-/**
- * Result from getSensors containing both individual sensors and averaged sensors
- */
-export interface SensorData {
-  /** Individual sensors specified in config.sensors */
-  individual: EntityState[];
-
-  /** Averaged sensors grouped by device class */
-  averaged: AveragedSensor[];
-}
-
-/**
- * Represents a display-ready sensor that can be either individual or averaged
- */
-export interface DisplaySensor {
-  /** Entity state (undefined for averaged sensors) */
-  state?: EntityState;
-
-  /** Display state value (undefined for individual sensors) */
-  value?: string;
-
-  /** Sensor domain */
-  domain: string;
-
-  /** Device class (e.g., 'temperature', 'humidity') */
-  device_class: string;
-}
