+++ test/cards/components/sensor-collection.spec.ts
@@ -6,7 +6,9 @@ import type { HomeAssistant } from '@hass/types';
 import * as stateDisplayModule from '@html/state-display';
 import { fixture } from '@open-wc/testing-helpers';
 import * as styleConverterModule from '@theme/util/style-converter';
-import type { Config, EntityState, SensorData } from '@type/config';
+import type { Config } from '@type/config';
+import type { EntityState } from '@type/room';
+import type { SensorData } from '@type/sensor';
 import { expect } from 'chai';
 import { html, nothing, type TemplateResult } from 'lit';
 import { stub } from 'sinon';
